requirements.txt
pexpect>=4.2.1
allennlp>=0.8.4
flair>=0.4.2
tf-sentencepiece>=0.1.82
tensorflow-hub>=0.5
docker>=4.0.2
git+git://github.com/oborchers/Fast_Sentence_Embeddings@34a6a33
git+git://github.com/facebookresearch/SentEval.git
psutil>=5.6.3
tensorflow-gpu==1.14.0